Modeling overhaul
The models for the enchant books have been heavily overhauled to allow for easier modification and customization.

Each now parent under the first iteration of the book to effectively have "one" model per book instead of separate ones for each individual book.

The textures themselves are now just emblems that get overlayed on top of a base, allowing for easier future texture creation/modification and for player customization.